Closed issues #7855, #7701, #7049, #8965

+++ db/migrate/031_change_kpi_period_users.rb
@@ -3,7 +3,7 @@ class ChangeKpiPeriodUsers < ActiveRecord::Migration
   def self.up
     change_table :kpi_period_users do |t|
       t.float :hours
-      t.integer :base_salary
+      t.decimal :base_salary
     end
   end
